Check2 updated
* userComputeapi interface and prototype.

created userCompute interface wit methods to set input, output and delimiters, also added the prototype to show an example of usage.

* storageCompute interface and prototype

Created storageComputeAPI interface with methods that read a list of inputs and writes formatted output, and the prototype demostrating how to read input and write the output

* ComputeEngineAPI interface and prototype

Created an interface with methods for single and multiple inputs, added a prototype showing an example for each.

* update code

contains request and response files

* updated code

* pass checkstyle

* Update StorageComputeAPIPrototype.java

* Update UserComputeAPIPrototype.java

* checkstyle

* updated code

the sum gets stored in a list and written to the output the user gives instead of just printing it. I also switched from using booleans to enums for status.

* fixed code so prototypes only contain 1 parameter